feat: add conditional aborting
